requirement.txt
Keras                 2.2.4
numpy                 1.16.4
tensorflow-gpu        1.13.1
